--- pkg/pare/cli/constants.py
@@ -1,3 +1,0 @@
-from __future__ import annotations
-
-GITHUB_CLIENT_ID = "Iv23li9dxAETy7AsXHzU"
