+++ pkg/pare/console.py
