+++ pkg/pare/constants.py
@@ -3,3 +3,5 @@ from __future__ import annotations
 import sys
 
 PYTHON_VERSION = f"{sys.version_info.major}.{sys.version_info.minor}"
+
+GITHUB_CLIENT_ID = "Iv23li9dxAETy7AsXHzU"
