+++ tach.toml
@@ -1,29 +1,42 @@
 exclude = [
     ".*__pycache__",
     ".*egg-info",
+    ".*tests",
     "api/src/transform/template.py",
     "docs",
-    ".*tests",
     "venv",
 ]
-source_roots = ["pkg"]
+source_roots = [
+    "pkg",
+]
 
 [[modules]]
 path = "pare"
-depends_on = [{ path = "pare.sdk" }]
+depends_on = [
+    { path = "pare.sdk" },
+]
 
 [[modules]]
 path = "pare.cli"
 depends_on = [
     { path = "pare.client" },
+    { path = "pare.console" },
     { path = "pare.constants" },
+    { path = "pare.login" },
     { path = "pare.models" },
     { path = "pare.settings" },
 ]
 
 [[modules]]
 path = "pare.client"
-depends_on = [{ path = "pare.settings" }]
+depends_on = [
+    { path = "pare.console" },
+    { path = "pare.settings" },
+]
+
+[[modules]]
+path = "pare.console"
+depends_on = []
 
 [[modules]]
 path = "pare.constants"
@@ -32,6 +45,14 @@ depends_on = []
 [[modules]]
 path = "pare.errors"
 depends_on = []
+
+[[modules]]
+path = "pare.login"
+depends_on = [
+    { path = "pare.console" },
+    { path = "pare.constants" },
+    { path = "pare.settings" },
+]
 
 [[modules]]
 path = "pare.models"
